Extract custom-palette runtime into theme-custom-palette.js
The custom-palette machinery — reading the stored v1/v2 payload, deriving the
runtime CSS custom properties via window.ThemeDerive, writing/clearing them on
<html>, and computing the dropdown swatch previews — was a self-contained island
inside theme.js (~260 lines) whose only outward dependency was the ThemeDerive
seam. Move it to its own module behind window.ThemeCustomPalette.

theme.js keeps four thin shims (has/applyStored/clearAppliedTokens/
syncOptionVisibility) so every call site stays unchanged and degrades to a
no-op if the module is absent, plus refreshCustomPaletteState as the palette
coordinator (it alone calls back into setPalette/updatePaletteUI, so it stays).
The seam is one-directional: the module consumes only ThemeDerive + the DOM.

Load theme-custom-palette.js before theme.js (after theme-derive.js) in both the
dev list and the prod concat; load it first in the tests that boot theme.js.
theme.js: 2174 -> 1939 lines. Full suite green (626).
