+++ assets/js/__tests__/theme-pantone.test.js
@@ -33,6 +33,9 @@ describe("darkmode pantone transport", () => {
 
   function loadModule() {
     jest.isolateModules(() => {
+      // theme-custom-palette.js publishes window.ThemeCustomPalette, which
+      // theme.js's custom-palette shims delegate to; load it first (as in prod).
+      require("../theme-custom-palette.js");
       require("../theme.js");
     });
     document.dispatchEvent(new window.Event("DOMContentLoaded"));
